Add the simplest IO bus in the world.
IO happens on page $FF because I say so, and right now the only IO
device is an output-only terminal that can be written to at $FF00.
Individual bytes written to $FF00 will be echoed to the terminal.

This also fixes a bug with bytecode interpretation where the PC was
being changed when it shouldn't be. The PC should always point to the
starting location of the current instruction or the location of the
next instruction, but I was moving it for every byte I read. This made
jumps do stupid things. Whoops.